# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Docling Serve WebUI API
  version: 1.0.0
  description: |
    WebUI backend API for batch image OCR processing via Docling Serve
    and LLM-powered table data extraction.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
paths:
...
  /files/{fileId}/llm:
    post:
      summary: Re-run LLM table extraction for a single file
      description: Re-processes the OCR result through LLM to extract table data. Useful for retrying failed LLM calls.
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                model:
                  type: string
                  description: Override LLM model for this request.
      responses:
        '202':
          description: LLM processing started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: File OCR not completed yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      summary: Get LLM result for a single file
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: LLM extraction result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMResultResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /files/{fileId}/export:
    get:
      summary: Export single file results as ZIP
      parameters:
        - $ref: '#/components/parameters/FileId'
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv, both]
            default: both
      responses:
        '200':
          description: ZIP file download
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /files/{fileId}/image:
    get:
      summary: Get original image file
      description: Serves the original uploaded image for preview in the browser.
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Image file
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
...
    FileId:
      name: fileId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: File UUID
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Error code (non-zero)
        message:
          type: string
          description: Human-readable error description
        details:
          type: object
          description: Additional error details
...
    FileDetailResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            batch_id:
              type: string
              format: uuid
            original_filename:
              type: string
            file_size:
              type: integer
            file_type:
              type: string
            ocr_status:
              type: string
              enum: [pending, processing, completed, failed]
            ocr_md_content:
              type: string
              nullable: true
              description: OCR result in Markdown format
            ocr_html_content:
              type: string
              nullable: true
              description: OCR result in HTML format
            ocr_processing_time:
              type: number
              nullable: true
            ocr_error:
              type: string
              nullable: true
            llm_status:
              type: string
              enum: [pending, processing, completed, failed, skipped]
            llm_result:
              type: object
              nullable: true
              description: LLM-extracted table data (parsed JSON object)
              properties:
                tables:
                  type: array
                  items:
                    type: object
                    properties:
                      table_index:
                        type: integer
                      headers:
                        type: array
                        items:
                          type: string
                      rows:
                        type: array
                        items:
                          type: object
            llm_model:
              type: string
              nullable: true
            llm_error:
              type: string
              nullable: true
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    LLMResultResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        data:
          type: object
          properties:
            file_id:
              type: string
              format: uuid
            llm_status:
              type: string
            llm_result:
              type: object
              nullable: true
              description: Parsed table data
            llm_model:
              type: string
              nullable: true
            llm_error:
              type: string
              nullable: true
...